add handbrake models
